haiku-generator/proof-of-concept/requirements-text-to-speech.txt
coqui-tts[server,ja]==0.25.3
torch==2.6.0
transformers==4.46.2